train/requirements.txt
torch==1.11.0
torchvision==0.12.0
pytorch_lightning==1.6.3
torchmetrics==0.8.2
numpy==1.21.6
scikit-learn==1.0.2
Pillow==7.1.2
matplotlib==3.2.2